api/requirements.txt
fastapi==0.115.6
google-auth==2.37.0
PyJWT==2.10.1
pydantic-settings==2.7.1
PyYAML==6.0.3
requests==2.32.3
SQLAlchemy==2.0.36
uvicorn[standard]==0.34.0
psutil==6.1.0
